requirements.txt
google-cloud-speech>=2.15.0
google-cloud-storage>=2.0.0 # Added for GCS uploads
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
numpy<2.0.0 # Downgraded to avoid 1.x vs 2.x conflict
pandas>=2.0.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
sounddevice>=0.4.6
wave>=0.0.2
websockets>=11.0.3
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
